Style filters form and move state to homepage to persist state between popover open and close
